Update README.md
In this project, we will analyze the French Open Damir database and build a machine learning 
regression model to predict the reimbursed amount of healthcare benefits. We will train four 
different models based on the concept of decision trees that use boosting or bagging or both. We will 
use grid search to tune their parameters. Finally, we will evaluate them and choose the best one 
according to specified metrics. The report includes an extensive analysis of the database, the social
security system as well as an introduction to ensemble learning and different machine learning 
techniques used